EmployeAssesment

+++ app/Database/Migrations/2025-04-28-065745_CreatePerformanceAssessmentTable.php
@@ -31,7 +31,7 @@ class CreatePerformanceAssessmentTable extends Migration
                 'unsigned'   => true,
             ],
             'nilai'   => [
-                'type'      => 'INT',
+                'type'      => 'FLOAT',
                 'constraint' => 11,
                 'null'      => true,
             ],
